Update on "[inductor] overlap scheduling nodes benchmarking logs and fix"
1. Fix: benchmark_collective_with_cuda_events_impl was benchmarking without convertion FakeTensor -> real
2. Adding logging to tlparse of runtime estimations 'benchmarked' vs 'analytical', will be logged only if 'benchmark' mode is on. (In this case overhead for analytical calculation is minimal).
3. As we do warmup=1, nruns=3, using median for stability, as warmup 1 can be not enough. median is more stable for outliers in this case.



cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx ipiszy kadeng muchulee8 amjames chauhang aakhundov coconutruben jataylo

[ghstack-poisoned]
